# specs/009-jev-decision-models/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: llmctl decision gateway
  version: 3.1.0-draft
  summary: Hosted-Jev-compatible typed-decision endpoint served by llmctl over HTTPS.
  description: |
    The decision gateway is the single network-facing decision endpoint of llmctl.
    It speaks the TypeSafe "System One" wire shape (`POST /v1/systemone`) so that clients
    built for the hosted service only need a different base URL, the access key, and trust
    for the generated certificate authority.

    Facts here marked [HOSTED] were established from the vendor documentation
    (docs.typesafe.ai, retrieved 2026-10-07; see specs/009-jev-decision-models/research/web-jev-api-and-bench.md).
    Facts marked [LLMCTL] are llmctl-defined because the vendor does not publish them.

    STATUS TABLE (single source; other documents cite this table):

    | Status | Condition | Retryable | Retry-After |
    |---|---|---|---|
    | 400 | malformed JSON, body not an object, wrong content type, more than 255 options [HOSTED answers 400 at 256] | no | no |
    | 401 | missing or wrong key | no | no |
    | 404 | unknown path (only after authentication) [LLMCTL] | no | no |
    | 405 | wrong method for the path [LLMCTL] | no | no |
    | 413 | body above the cap, rejected before reading [LLMCTL] | no | no |
    | 422 | schema or limit violation, state over budget, unknown model, deterministic readout failure (readout_failed) | no | no |
...

...

...
    loopback-only and are NOT part of this contract.
servers:
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
  /v1/systemone:
    post:
      operationId: systemOne
      summary: Ask typed questions about a state.
      description: |
        One request may carry several independent questions keyed by caller-chosen names
        [HOSTED]. Names only key the answers and never reach the model.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SystemOneRequest" }
      responses:
        "200":
          description: Typed answers keyed by question name.
          headers:
            x-llmctl-request-id: { $ref: "#/components/headers/RequestId" }
            x-llmctl-decide-mode:
              description: "[LLMCTL] The serving mode of the gateway: `deterministic` (fixed seed, one slot per instance, byte-identical repeats per instance) or `throughput` (least-loaded instances, byte-identity across instances NOT promised). Present on every decision."
              schema: { type: string, enum: [deterministic, throughput] }
            x-llmctl-decide-instance:
              description: "[LLMCTL] The engine instance that answered (for example `decide-tiny` or `decide-tiny.2`). Deterministic mode promises byte-identical repeats PER INSTANCE and a busy primary overflows to the next instance (FR-074), so two answers are only comparable when this value is equal. Present on every decision whose backend reports one; hosted SDKs do not surface custom headers."
              schema: { type: string, pattern: "^[a-z0-9._-]{1,64}$" }
            x-llmctl-decide-truncated:
              description: "[LLMCTL] Present and true ONLY when the operator enabled the opt-in state-shortening setting and the state was shortened (head and tail kept by the gateway, or the premise shortened by an encoder runtime). The default is to reject an over-long state with 422. Question and option text are never shortened. Hosted SDKs do not surface custom headers."
              schema: { type: string, enum: ["true"] }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SystemOneResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "413": { $ref: "#/components/responses/TooLarge" }
        "422": { $ref: "#/components/responses/Unprocessable" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/Misconfigured" }
        "502": { $ref: "#/components/responses/BackendFailure" }
        "503": { $ref: "#/components/responses/NotReady" }
        "529": { $ref: "#/components/responses/Overloaded" }
  /v1/models:
    get:
      operationId: listModels
      summary: List served decision models and accepted aliases.
      description: "[LLMCTL] The vendor SDKs (typesafe-sdk 0.7.2, @typesafe-ai/sdk 0.6.0) read this endpoint as {models:[{name,description,release_date}]} and reject any other shape, so the body carries that SDK shape (`models`) AND the llmctl-defined listing (`object`/`data`) side by side."
      responses:
        "200":
          description: Served models.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModelList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "503": { $ref: "#/components/responses/NotReady" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "`Authorization: Bearer $LLMCTL_API_KEY`. The key is never accepted in the query string."
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Alias for clients that cannot set Authorization.
  headers:
    RequestId:
      description: "[LLMCTL] Correlation identifier (random 64-bit hex) echoed in logs. The vendor uses x-typesafe-request-id [HOSTED]."
      schema: { type: string, pattern: "^[0-9a-f]{16}$" }
    RetryAfter:
      description: Seconds to wait before retrying.
      schema: { type: integer, minimum: 0 }
  schemas:
    SystemOneRequest:
      type: object
      required: [state, questions]
      additionalProperties: false
      properties:
        model:
          type: string
          description: |
            Accepted: `jev-latest`, `jev-preview`, `jev-1.13.0` (aliases mapped to the default served profile),
            `llmctl-<profile>` and `<profile>`. The answer reports the REAL local model id. Unknown name -> 422 `unknown_model`.
        state:
          description: Text, an object, or an array. State plus the longest question must fit the profile's budget.
          oneOf:
            - { type: string }
            - { type: object }
            - { type: array }
        questions:
          type: object
          minProperties: 1
          maxProperties: 32
          description: "Keys are free names. The maximum count is a configurable llmctl limit (default 32)."
          additionalProperties:
            $ref: "#/components/schemas/Question"
    Question:
      oneOf:
        - $ref: "#/components/schemas/NoulQuestion"
        - $ref: "#/components/schemas/ChoiceQuestion"
        - $ref: "#/components/schemas/ScoreQuestion"
      discriminator:
        propertyName: type
...
    SystemOneResponse:
      type: object
      required: [model, answers, usage]
      additionalProperties: false
      properties:
        model: { type: string, description: "The real local model identifier." }
        answers:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/Answer"
        usage:
          type: object
          required: [input_tokens, output_tokens]
          additionalProperties: false
          properties:
            input_tokens: { type: integer, minimum: 0, description: "[LLMCTL] An estimate (about characters/4) unless the backend reports exact counts." }
            output_tokens: { type: integer, minimum: 0 }
    Answer:
      oneOf:
        - $ref: "#/components/schemas/NoulAnswer"
        - $ref: "#/components/schemas/ChoiceAnswer"
        - $ref: "#/components/schemas/ScoreAnswer"
      discriminator:
        propertyName: type
...
    ModelList:
      type: object
      required: [object, data, models]
      additionalProperties: false
      properties:
        models:
          type: array
          description: "[HOSTED SDK SHAPE] One entry per accepted model name: each served profile id and each of its aliases (the value of the request `model` field)."
          items:
            type: object
            required: [name, description, release_date]
            additionalProperties: false
            properties:
              name: { type: string, description: "Profile id or alias accepted by the request model field." }
              description: { type: string, description: "Catalog `desc` of the profile (an alias entry is prefixed `Alias of <id>.`)." }
              release_date:
                type: string
                format: date
                description: "YYYY-MM-DD. The profile's `decision.release_date` when the catalog pins one, else the fixed documented default 2026-10-07 (the publication date of this contract). Never stamped per request."
        object: { const: list }
        data:
          type: array
          items:
            type: object
            required: [id, aliases, protocol, status, limits]
            additionalProperties: false
            properties:
              id: { type: string, description: "Local profile id." }
              aliases: { type: array, items: { type: string } }
              protocol: { type: string, enum: [letter-logit, systemone-native, nli-onnx] }
              status: { type: string, enum: [ready, starting, degraded, draining] }
              limits:
                type: object
                required: [max_options, score_levels]
                additionalProperties: false
                description: "[LLMCTL] Per-profile limits; the hosted maximum of 255 options applies only where the profile supports it."
                properties:
                  max_options: { type: integer, minimum: 2, maximum: 255 }
                  max_state_chars: { type: integer, minimum: 1, description: "[LLMCTL] The largest ordinary-prose state (characters) the profile accepts: the configured character cap lowered to what the engine context holds. Acceptance is decided by an ESTIMATE in tokens (digits, JSON and non-Latin scripts cost more per character), so a denser state of fewer characters can be refused with 422 validation_failed." }
                  max_context_tokens: { type: integer, minimum: 1, description: "[LLMCTL] Estimated tokens the rendered prompt may use: the catalog ctx (per request: llama-server divides --ctx-size over its slots and the scheduler multiplies it back) minus the chat-template reserve and the reply. Encoder profiles: the pair window (premise + hypothesis)." }
                  max_pairs: { type: integer, minimum: 1, description: "[LLMCTL] Encoder (nli-onnx) profiles: the most premise/hypothesis pairs (questions x options) one request may cost; over it is a 422 before any pass is spent." }
                  score_levels:
                    type: array
                    items: { type: integer }
                    minItems: 2
                    maxItems: 2
              notes: { type: string, description: "Honest labels, e.g. experimental choice/score on the NLI profile." }
              template_hash:
                type: string
                pattern: "^[0-9a-f]{64}$"
                description: "[LLMCTL] The profile's `decision.template_hash`: SHA-256 of the prompt template (as the gateway renders it) and of the readout parameters that move the probabilities (spellings, n_probs, the global readout temperature). Computed by the gateway (never stored in the catalog); calibration profiles are bound to it. Absent if it cannot be computed."
              calibration:
                type: object
                required: [applied]
                additionalProperties: false
                description: "[LLMCTL] Present only when a calibration profile file exists for the model (`$STATE/decide/calibration/<profile>.json`). Applied, or - with a closed `reason` - not applied; a refused profile leaves answers uncalibrated."
                properties:
                  applied: { type: boolean }
                  method: { type: string, enum: [temperature, platt, isotonic], description: "Applied only." }
                  n: { type: integer, minimum: 0, description: "Labelled answers the profile was fitted on (applied only). Under 200 supports no ECE claim (SC-003)." }
                  profile_id: { type: string, description: "The calibration profile name (applied only)." }
                  reason: { type: string, enum: [mismatch, unbound, invalid, insecure, model_unresolved], description: "Not applied only. mismatch = bound to another model file, prompt template or readout temperature; unbound = not bound; invalid = unreadable or malformed parameters; insecure = symlink, not a regular file or writable by group/others; model_unresolved = the catalog names no single model sha256." }
...
    Error:
      type: object
      required: [message, error_type]
      additionalProperties: false
      properties:
        message: { type: string, description: "Generic text; never exception text, paths or versions." }
        error_type:
          type: string
          description: "[LLMCTL] The vendor's full vocabulary is not published (only `invalid_request` is documented)."
          enum:
            - invalid_request
            - unauthorized
            - unknown_model
            - validation_failed
            - payload_too_large
            - method_not_allowed
            - rate_limited
            - overloaded
            - not_ready
            - readout_failed
            - backend_failed
  responses:
    BadRequest:
      description: "Malformed JSON, non-object body, wrong content type, or more than 255 options in a choice question (the hosted service answers 400 at 256 options)."
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unauthorized:
      description: Missing or wrong key. Generic body.
      headers:
        WWW-Authenticate: { schema: { type: string, const: 'Bearer realm="llmctl"' } }
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    MethodNotAllowed:
      description: Wrong method for the path (after authentication).
      headers:
        Allow: { schema: { type: string } }
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    TooLarge:
      description: Body above the configured cap, rejected before it is read.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unprocessable:
      description: "Schema or limit violation (option and level counts above the profile cap, state over the profile budget, unknown model) and deterministic readout failure (readout_failed: the letter-probability mass was below the threshold; the model would answer the same way again, so it is not retryable)."
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    RateLimited:
      description: Failed-authentication or per-source throttle.
      headers:
        Retry-After: { $ref: "#/components/headers/RetryAfter" }
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Overloaded:
      description: Capacity saturated (all slots busy and the bounded queue is full or its deadline passed). Retryable.
      headers:
        Retry-After: { $ref: "#/components/headers/RetryAfter" }
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotReady:
      description: No ready instance, or the gateway is draining.
      headers:
        Retry-After: { $ref: "#/components/headers/RetryAfter" }
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    BackendFailure:
      description: "Transient backend failure (process died, connection refused). Retryable. No fabricated probabilities are ever returned."
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Misconfigured:
      description: "[LLMCTL] Deterministic server-side fault: a misconfigured readout setting, or an engine that rejects the gateway's own request shape. NOT retryable (a retry fails the same way); the cause is in the gateway log. A request the model cannot fit (state too long for the engine context, empty state for an encoder, over the pair budget) is a 422 validation_failed instead, never this."
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }